modules/Assignment Aider/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.6
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
filelock==3.16.1
Flask==3.0.3
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.66.1
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
idna==3.10
imageio==2.35.1
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langchain==0.3.0
langchain-core==0.3.5
langchain-google-genai==2.0.0
langchain-text-splitters==0.3.0
langsmith==0.1.128
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mistune==3.0.2
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.1.0
narwhals==1.8.3
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
opencv-python-headless==4.10.0.84
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pdf2image==1.17.0
pillow==10.4.0
proto-plus==1.24.0
protobuf==4.25.5
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAutoGUI==0.9.54
pyclipper==1.3.0.post5
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
pyparsing==3.1.4
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-xlib==0.15
pytweening==1.2.0
pytz==2024.2
pywhatkit==5.4
PyYAML==6.0.2
referencing==0.35.1
reportlab==4.2.2
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
rsa==4.9
scikit-image==0.24.0
scipy==1.14.1
setuptools==75.1.0
shapely==2.0.6
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
streamlit==1.38.0
sympy==1.13.3
tenacity==8.5.0
tifffile==2024.9.20
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
watchdog==4.0.2
Werkzeug==3.0.4
wikipedia==1.4.0
yarl==1.12.1
